SurfaceFlinger: Set debug.sf.frame_rate_multiple_threshold to 60 by default
If the device only have 60 and 120 Hz in the config group, 60 Hz won't be
used when watching a 24 fps video and instead 120 Hz will be used because
it is perfectly divisible by 24.

This is a publicly documented AOSP behavior [1], but it simply does not
make any practical sense and trades off significant amount of power for a
very-hard-to-notice jitter-free experience.

Set debug.sf.frame_rate_multiple_threshold to 60 by default to limit this
behavior to 60 or lower fps.

[1] https://source.android.com/devices/graphics/multiple-refresh-rate

Change-Id: I66353abc46ad16e63b2c9188794469ad71f2192e
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>